fuelquote app so far
got the urls and views to see the webpage done, was trying to work on the models.py after doing the codeacademy but i was still confused, im pretty sure i need to add a foreign key for the delivery address but still need to figure that out. Still need to do the  includes list of quote history for a client. portion of this module, trying to look at thing right now to see whats the best way to approach it